requirements.txt
# Core dependencies for Home Assistant
PyYAML>=6.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
# pyaudio>=0.2.11  # Requires PortAudio system dependency

# Optional dependencies for advanced speech recognition
# google-cloud-speech  # For Google Speech-to-Text
# openai-whisper      # For OpenAI Whisper

# Development and testing dependencies
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0